Added image_pair_plot as a file and added DOP calculations. PEP8 corrections here and there.
